requirements.txt
# Bloomberg Data Crawler Dependencies

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pandas>=2.0.0

# API clients
yfinance>=0.2.32
websocket-client>=1.6.0
aiohttp>=3.9.0
aiolimiter>=1.1.0

# Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Scheduling
apscheduler>=3.10.0

# Data serialization
orjson>=3.9.0

# Logging
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.23.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
